requirements.txt
gradio>=4.0.0
openai>=1.0.0
pinecone>=2.2.4
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
tiktoken>=0.5.0 